refactor: move field logic out of RichTextEditor
